refactor: simplify parse_default_mode return
parse default will now return a list of dataclass
